requirements.txt
pinecone-client==2.2.4
streamlit==1.29.0
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.23
